+++ app/Services/MetricsCollectionService.php
@@ -377,11 +377,11 @@ class MetricsCollectionService
         }
 
         return sprintf(
-            'ssh %s %s@%s "%s"',
+            'ssh %s %s@%s %s',
             implode(' ', $sshOptions),
             $server->username,
             $server->ip_address,
-            addslashes($remoteCommand)
+            escapeshellarg($remoteCommand)
         );
     }