+++ app/Services/ServerBackupService.php
@@ -468,12 +468,12 @@ BASH;
             $escapedPassword = escapeshellarg($server->ssh_password);
 
             return sprintf(
-                'sshpass -p %s ssh %s %s@%s "%s"',
+                'sshpass -p %s ssh %s %s@%s %s',
                 $escapedPassword,
                 implode(' ', $sshOptions),
                 $server->username,
                 $server->ip_address,
-                addslashes($command)
+                escapeshellarg($command)
             );
         }
 
@@ -487,11 +487,11 @@ BASH;
         }
 
         return sprintf(
-            'ssh %s %s@%s "%s"',
+            'ssh %s %s@%s %s',
             implode(' ', $sshOptions),
             $server->username,
             $server->ip_address,
-            addslashes($command)
+            escapeshellarg($command)
         );
     }